Update monitor to use floats

+++ tamr_toolbox/utils/operation.py
@@ -129,7 +129,7 @@ def wait(
 
 
 def monitor(
-    operation: Operation, *, poll_interval_seconds: int = 1, timeout_seconds: int = 300,
+    operation: Operation, *, poll_interval_seconds: float = 1, timeout_seconds: float = 300,
 ) -> Operation:
     """Continuously polls for this operation's server-side state and returns operation
     when there is a state change
